partial fix of issue #5, move cobra_edit_glossary_form class into its own file
