+++ lib/glossarylib.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * External cobra API
+ * CoBRA functions dedicated to glossary management.
  *
  * @package    mod_cobra
  * @author     Jean-Roch Meurisse
@@ -182,37 +182,4 @@ function cobra_mark_unknown_word($word, $text) {
     $text = str_replace(' ' . $word . ')', ' <span style="color:red">' . $word . '</span>)', $text);
     $text = str_replace('(' . $word . ' ', '(<span style="color:red">' . $word . '</span> ', $text);
     return $text;
-}
-
-/**
- * Class cobra_edit_glossary_form
- *
- * @package    mod_cobra
- * @author     Laurence Dumortier
- * @copyright  2016 onwards - Cellule TICE - Universite de Namur
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-class cobra_edit_glossary_form extends moodleform {
-    /**
-     * Form definition.
-     *
-     * @throws coding_exception
-     */
-    public function definition() {
-        $mform = $this->_form;
-        // 1st argument is group name, 2nd is link text, 3rd is attributes and 4th is original value.
-        $this->add_checkbox_controller(1, null, null, 1);
-        $textlist = $this->_customdata['textlist'];
-        $compare = $this->_customdata['compare'];
-        foreach ($textlist as $text) {
-            $mform->addElement('advcheckbox', 'text_' . $text->id, '',
-                    htmlspecialchars(strip_tags($text->name)), array('group' => 1));
-            $mform->setDefault('text_' . $text->id, 1);
-
-        }
-        if ($compare) {
-            $mform->addElement('textarea', 'mytext', get_string('mytext', 'cobra'), array('rows' => 30, 'cols' => 80));
-        }
-        $this->add_action_buttons(true, get_string('OK', 'cobra'));
-    }
 }
